Add support for using benmills/vimux
# What
Add a global config option for enabling sending dlv commands to a nearby
tmux pane using benmills/vimux, instead of sending commands to a
terminal buffer within vim.

# Why
For those of us with a workflow where we run tests outside of vim using
vimux and other plugins like vim-test, it better integrates delve more
seamlessly. Instead of running tests in an adjacent tmux pane, then
running delve inside vim, both can be run in an adjacent tmux pane.

# Notes
The changes also include:

1. Not setting the `--backend` parameter if it's value is `default`.
Normally users of this plugin do not see the command that is executed,
but when using vimux they do. It's unnecessary to set the value of that
parameter with default as that is the default value if not set, so it
reduces the length of the command.

2. Adding a `cd -` to the end, since a tmux pane will be reused and it's
nice if changing the directory if we return the user to their original
directory.