Add function to get init instructions
# What
Add function to get init instructions.

# Why
The init instructions that the vim-delve plugin builds up that contains
breakpoints, tracepoints, etc are useful to other plugins, but they are
inaccessible because they are store in a script-scoped variable. Adding
this function to expose the instructions allows other plugins to tap
into the breakpoints, tracepoints, etc that the vim-delve plugin are
tracking for the user.