override file path for breakpoint and tracepoint
If the Go executable was built in a different file path than where you
are editing the files, then setting breakpoints and/or tracepoints will
set the filepath of the file where you have it opened, since delve
doesn't know the location that the plugin is transmitting when using
DlvConnect. With this commit you are able to set the project root path
that vim-delve will use when transmitting info to delve about breakpoints
and/or tracepoints.